Add initial Docker and Docker Compose set up
Adds an initial set up for Docker and Docker Compose. The assumption
is that we will build a "local" image for use with Docker Compose, so
a `docker` folder has been added with a script and `Dockerfile` needed
to build the node image.

README is updated with basic instructions to get up and running.