Add direct color input to color picker
Adds the ability to directly enter the color in the color picker
component by converting the previous read-only display to an input. The
`color` property is bound to the input and a new function is added to
allow the input to emit the color so the other components can respond
accordingly.

The input itself does not do any validation. The input is validated on
save in the higher level component where these are used.